refactor: split client provider DOM helpers
